move datadog metrics from app svc-bip-api into a new shared lib namespace of the new lib: gov.va.vro.metricslogging
