+++ settings.gradle
@@ -36,6 +36,7 @@ include ':dev-tools'
 include ':postgres'
 include ':shared:persistence-model'
 include ':shared:lib-bie-kafka'
+include ':shared:lib-metrics'
 include ':shared:lib-rabbitmq-connector'
 include ':svc-bgs-api'
 include ':svc-bip-api'
@@ -51,3 +52,4 @@ include ':domain-cc:cc-app'
 include ':domain-ee:ee-max-cfi-app'
 include ':domain-ee:ee-ep-merge-app'
 include 'rabbitmq'
+
