+++ svc-bip-api/build.gradle
@@ -11,6 +11,7 @@ ext {
 
 dependencies {
   implementation project(':domain-xample:xample-shared')
+  implementation project(":shared:lib-metrics")
 
   implementation "org.springdoc:springdoc-openapi-starter-webmvc-api:${spring_doc_version}"
   implementation 'org.apache.httpcomponents:httpclient:4.5.14'
